refactor(macro-commons): promote WiringGraph + WiringGraphLogging to the shared module
Move the ZIO-Magic-style wiring-graph model/renderers out of di's internals
into kindlings-macro-commons (as private[kindlings]) and factor the
graph-logging boilerplate (mode toggle + emit helpers) into a shared
WiringGraphLogging trait, so both di and di-cats can reuse them. No behaviour
change for di.
